don't destroy the backlog if it's < 140 characters
With 1.2.0, get_new_zombies returns each name exactly once(unless the
person is cured and killed again). In prior versions names would be
returned until they'd been reported. The backlog was only kept if it was
at least one tweet worth. In 1.2.0, if more than 140 characters were
reported, but not enough to fill a second 140, names would be lost.